requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
requests==2.31.0
aiohttp==3.9.1
asyncio
uuid
logging
urllib3==2.1.0